feat: add new forgotten password page with updated styling
Implement forgotten password form matching the new login page design.
Extract shared AuthPageLayout component for the page shell (background,
logo header, card with decorative squares). Update form gap to 24px
for both login and forgotten password forms.
