+++ frontend/src/component/user/ForgottenPassword/DeprecatedForgottenPassword.tsx
@@ -1,33 +1,27 @@
-import {
-    Button,
-    Divider,
-    styled,
-    TextField,
-    Typography,
-    AlertTitle,
-    Alert,
-} from '@mui/material';
+import { Button, styled, TextField, Typography } from '@mui/material';
+import { AlertTitle, Alert } from '@mui/material';
 import { type SyntheticEvent, useState } from 'react';
 import { Link } from 'react-router-dom';
 import useLoading from 'hooks/useLoading';
 import { FORGOTTEN_PASSWORD_FIELD } from 'utils/testIds';
 import { formatApiPath } from 'utils/formatPath';
 import { ConditionallyRender } from 'component/common/ConditionallyRender/ConditionallyRender';
-import { useFlag } from '@unleash/proxy-client-react';
-import DeprecatedForgottenPassword from './DeprecatedForgottenPassword';
-import { AuthPageLayout } from '../common/AuthPageLayout';
+import DividerText from 'component/common/DividerText/DividerText';
+import StandaloneLayout from '../common/StandaloneLayout.tsx';
+import {
+    contentSpacingY,
+    flexColumn,
+    textCenter,
+    title,
+} from 'themes/themeStyles';
 
-const StyledTitle = styled(Typography)(({ theme }) => ({
-    fontSize: theme.typography.h2.fontSize,
-    fontWeight: theme.typography.fontWeightBold,
-    lineHeight: '28px',
-    marginBottom: theme.spacing(3),
-}));
-
-const StyledForm = styled('form')(({ theme }) => ({
-    display: 'flex',
-    flexDirection: 'column',
-    gap: theme.spacing(3),
+const StyledDiv = styled('div')(({ theme }) => ({
+    ...contentSpacingY,
+    ...flexColumn,
+    width: '350px',
+    [theme.breakpoints.down('sm')]: {
+        width: '100%',
+    },
 }));
 
 const StyledStrong = styled('strong')(({ theme }) => ({
@@ -35,15 +29,34 @@ const StyledStrong = styled('strong')(({ theme }) => ({
     margin: theme.spacing(1, 0),
 }));
 
+const StyledButton = styled(Button)(({ theme }) => ({
+    width: '150px',
+    margin: theme.spacing(2, 'auto'),
+}));
+
+const StyledTitle = styled('h2')(({ theme }) => ({
+    ...title(theme),
+    ...textCenter,
+}));
+
+const StyledForm = styled('form')(({ theme }) => ({
+    ...contentSpacingY(theme),
+    ...flexColumn,
+}));
+
+const StyledTypography = styled(Typography)(({ theme }) => ({
+    ...textCenter,
+}));
+
 type State = 'initial' | 'loading' | 'attempted' | 'too_many_attempts';
 
-const NewForgottenPassword = () => {
+const DeprecatedForgottenPassword = () => {
     const [email, setEmail] = useState('');
     const [state, setState] = useState<State>('initial');
     const [attemptedEmail, setAttemptedEmail] = useState('');
     const ref = useLoading(state === 'loading');
 
-    const onSubmit = async (e: SyntheticEvent) => {
+    const onClick = async (e: SyntheticEvent) => {
         e.preventDefault();
         setState('loading');
         setAttemptedEmail(email);
@@ -64,13 +77,13 @@ const NewForgottenPassword = () => {
     };
 
     return (
-        <AuthPageLayout>
-            <div ref={ref}>
-                <StyledTitle>Forgot your password?</StyledTitle>
+        <StandaloneLayout>
+            <StyledDiv ref={ref}>
+                <StyledTitle data-loading>Forgotten password</StyledTitle>
                 <ConditionallyRender
                     condition={state === 'attempted'}
                     show={
-                        <Alert severity='success' sx={{ mb: 3 }}>
+                        <Alert severity='success' data-loading>
                             <AlertTitle>Attempted to send email</AlertTitle>
                             We've attempted to send a reset password email to:
                             <StyledStrong>{attemptedEmail}</StyledStrong>
@@ -84,7 +97,7 @@ const NewForgottenPassword = () => {
                 <ConditionallyRender
                     condition={state === 'too_many_attempts'}
                     show={
-                        <Alert severity='warning' sx={{ mb: 3 }}>
+                        <Alert severity='warning' data-loading>
                             <AlertTitle>
                                 Too many password reset attempts
                             </AlertTitle>
@@ -92,53 +105,55 @@ const NewForgottenPassword = () => {
                         </Alert>
                     }
                 />
-                <StyledForm onSubmit={onSubmit}>
-                    <Typography variant='body2'>
-                        Enter your email address. If it exists in the system,
-                        we'll send you a reset password link.
-                    </Typography>
+                <StyledForm onSubmit={onClick}>
+                    <StyledTypography variant='body1' data-loading>
+                        Please provide your email address. If it exists in the
+                        system we'll send a new reset link.
+                    </StyledTypography>
                     <TextField
-                        label='Email'
                         variant='outlined'
                         size='small'
+                        placeholder='email'
                         type='email'
+                        data-loading
                         data-testid={FORGOTTEN_PASSWORD_FIELD}
                         value={email}
                         onChange={(e) => {
                             setEmail(e.target.value);
                         }}
-                        autoFocus
                     />
-                    <Button
+                    <StyledButton
                         variant='contained'
                         type='submit'
+                        data-loading
                         color='primary'
                         disabled={state === 'loading'}
-                        fullWidth
                     >
-                        Reset password
-                    </Button>
-                    <Divider>OR</Divider>
+                        <ConditionallyRender
+                            condition={state === 'initial'}
+                            show={<span>Submit</span>}
+                            elseShow={<span>Try again</span>}
+                        />
+                    </StyledButton>
+                    <DividerText text='Or log in' />
                     <Button
+                        type='submit'
+                        data-loading
                         variant='outlined'
+                        disabled={state === 'loading'}
                         component={Link}
                         to='/login'
-                        fullWidth
+                        sx={(theme) => ({
+                            width: '150px',
+                            margin: theme.spacing(2, 'auto'),
+                        })}
                     >
-                        Go back to sign in
+                        Log in
                     </Button>
                 </StyledForm>
-            </div>
-        </AuthPageLayout>
+            </StyledDiv>
+        </StandaloneLayout>
     );
 };
 
-const ForgottenPassword = () => {
-    const newLogin = useFlag('newLogin');
-    if (newLogin) {
-        return <NewForgottenPassword />;
-    }
-    return <DeprecatedForgottenPassword />;
-};
-
-export default ForgottenPassword;
+export default DeprecatedForgottenPassword;
